requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.2.7
celery-singleton==0.3.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
cookiecutter==2.6.0
Django==5.0.4
django-browser-reload==1.12.1
django-cachalot==2.6.2
django-cors-headers==4.3.1
django-redis==5.2.0
flower==2.0.1
humanize==4.9.0
idna==3.7
Jinja2==3.1.3
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tornado==6.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
